chore: Moves Collapse to the components folder (#14124)
