+++ superset-frontend/src/explore/components/DataTablesPane/index.tsx
@@ -18,7 +18,7 @@
  */
 import React, { useCallback, useEffect, useState } from 'react';
 import { JsonObject, styled, t } from '@superset-ui/core';
-import Collapse from 'src/common/components/Collapse';
+import Collapse from 'src/components/Collapse';
 import Tabs from 'src/common/components/Tabs';
 import Loading from 'src/components/Loading';
 import TableView, { EmptyWrapperType } from 'src/components/TableView';
