+++ superset-frontend/src/explore/components/DatasourcePanel/index.tsx
@@ -18,7 +18,7 @@
  */
 import React, { useEffect, useState } from 'react';
 import { styled, t } from '@superset-ui/core';
-import Collapse from 'src/common/components/Collapse';
+import Collapse from 'src/components/Collapse';
 import {
   ColumnOption,
   MetricOption,
